+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+- Feat: enable extra plugins in development mode
+
 ## [0.1.2] - 2022-05-30
 
 - Fix: use UTF-8 encoding for file reads/writes